Add files via upload
A library management system keeps track of the books present in the library. It is an important piece of software which is a must at schools and colleges.

Python is the most popular programming language on the planet. Python is an object-oriented language, which means that it can model real-world entities. It is dynamically-typed because it carries out type-checking at runtime.